{topost}: net: hns: add new architecture of dsaf
For 4.5 kernel, the dts define one mbigen node with a device id,
but the old architecture have 4 interrupt device id, which is not
compatible with the new mbigen driver. So we devide the dasf node
into 3 independnet dsaf0, dsaf1, dsaf2.

dsaf0 is for service port, dsaf1 is for debug port0, dsaf1 is for
debug port1.

Signed-off-by: Daode Huang <huangdaode@hisilicon.com>
Signed-off-by: Yisen Zhuang <yisen.zhuang@huawei.com>