ACPI: add ACPI_DEV to linux/acpi.h
dev of the acpi_device struct will be used for drivers, it will
be compiled in non-ACPI case, but acpi_device is in acpi_bus.h
and it can only be used in ACPI case. if drivers have the code to
access dev of acpi_device, it will be failed to compiled in non-ACPI
case. so this patch adds ACPI_DEV to return the dev pointer in
acpi_device in ACPI case while returns NULL in the other cases.

Signed-off-by: Kejian Yan <yankejian@huawei.com>
cc: Rafael <rjw@rjwysocki.net>