net: hns: register phy device in each mac initial sequence
In ACPI situation, There is no interface for registering phy device to
mdio-bus. Phy device has to be registered itself to mdio-bus, and them
enet can get the phy device's info so that it can config the phy-device
to help to trasmit and receive data.
Phy device is registered in each mac initial sequence. It depands on the
info on Hns Dsdt table in acpi coding by asl.

The hns dsdt table sample as below:
Scope(_SB)
{
        ...
        Device (MDIO)
        {
                ...
        }

        Device (DSF0)
        {
                ...

                Device (PRT0)
                {
                        Name (_ADR, 0x0)
                        Name (_DSD, Package () {
                                ...
                                Package () {
                                                Package () {"port-id", 0},
                                                Package () {"phy-mode", "sgmii"},
                                                Package () {"phy-addr", 0},
                                                Package () {"mdio-node", Package (){\_SB.MDIO}},
                                        }
                                ...
                       })
                }

                ...
        }

        ...
}

Signed-off-by: Kejian Yan <yankejian@huawei.com>