hisi_sas: fix v2 hw multiple SATA disk issue
Intermittently it is found that when multiple SATA
disks are directly connected to the host that some
disks are not detected.
The problem is that all set bitfields in ENT_INT_SRC1
are cleared for all phys in sata_int_v2_hw() - it
should clear the set bit for the phy being serviced.

Also unnecessary double-write to ENT_INT_SRC1 and
ENT_INT_SRC_MSK1 is removed (remaining writes are
done at end label).

Signed-off-by: John Garry <john.garry@huawei.com>