net: hns: use device_get_mac_address()
Hns cannot get mac address in initial sequence. It shows the log as below:
 hns-nic HISI00C1:00: No valid mac, use random mac 92:cd:d6:ec:74:61
 hns-nic HISI00C1:01: No valid mac, use random mac ae:ef:80:95:fd:ff
 hns-nic HISI00C1:02: No valid mac, use random mac b6:f0:bd:4e:14:87
 hns-nic HISI00C1:03: No valid mac, use random mac f6:f3:d9:24:f4:4b

It cause by misusing function to get mac address.of_get_mac_address()
is used to get mac address only in OF case, so we use
device_get_mac_address() instead. It can be used in both acpi and OF cases

Signed-off-by: Kejian Yan <yankejian@huawei.com>